feat: Migrate song stats dashboard to a multi-page app
